It's landing successfully sometimes!
Adding batchSize setting for multiple tests of one set of parameters in
different environments

Increasing sensor limits for gentler curves

Removing actor limits as they do NOT need to be re-scaled. When learning
properly the weights should end up returning outputs in the needed
range. Attempting to scale/denormalize the outputs was a majorly bad
idea. Kind of what I get for blindly listening to lectures :)

Unsquaring error. Professor suggested squaring it, and I'll try again,
but for now it's landing successfully unsquared. One thing at a time.

Returning reward of 1 for landing. If it lands, GOOD. Don't care how
well it landed.